refactor: move shared types closer to uses
