--- src/components/filters/filters.types.ts
@@ -1,10 +1,0 @@
-export type DomainValue = {
-  name: string;
-  code: string;
-};
-
-export type SpeciesLengthRow = {
-  species: string;
-  min: string;
-  max: string;
-};
